enabled ts support for the project
