+++ package.json
@@ -1,20 +1,26 @@
 {
-	"name": "csv-to-sql-insert",
-	"version": "0.1.0",
-	"type": "module",
-	"repository": "https://github.com/gitdagray/csv-to-sql",
-	"description": "input a csv file, output a sql insert statement",
-	"main": "index.js",
-	"author": {
-		"name": "Dave Gray",
-		"email": "dave@davegray.codes",
-		"url": "https://www.davegray.codes/"
-	},
-	"license": "MIT",
+  "name": "csv-to-sql-insert",
+  "version": "0.1.0",
+  "type": "module",
+  "repository": "https://github.com/gitdagray/csv-to-sql",
+  "description": "input a csv file, output a sql insert statement",
+  "main": "index.js",
+  "scripts": {
+    "start": "npm run build && node dist/index.js",
+    "build": "tsc  --skipLibCheck"
+  },
+  "author": {
+    "name": "Dave Gray",
+    "email": "dave@davegray.codes",
+    "url": "https://www.davegray.codes/"
+  },
+  "license": "MIT",
 	"devDependencies": {
 		"husky": "^8.0.3",
 		"prettier": "3.1.1",
-		"pretty-quick": "^3.1.3"
+		"pretty-quick": "^3.1.3",
+		 "@types/node": "^20.10.5",
+    "typescript": "^5.3.3"
 	},
 	"husky": {
 		"hooks": {
